Yhdistetty selectCylinderSetsByOwner ja selectCylinderSets.
